Clean up controllers
1. Move all ControllerInterface implementations to [framework]_controller.go
2. Clean up job.go pod.go service.go status.go
3. Follow exact same method order here https://github.com/kubeflow/common/blob/9ffa565bc60e08936f7f80cb3f491cf53f256e7f/pkg/apis/common/v1/interface.go
