+++ examples/xgboost/xgboostjob.yaml
@@ -10,7 +10,7 @@ spec:
       template:
         spec:
           containers:
-          - name: xgboostjob
+          - name: xgboost
             image: docker.io/merlintang/xgboost-dist-iris:1.1
             ports:
             - containerPort: 9991
@@ -29,7 +29,7 @@ spec:
       template:
         spec:
           containers:
-          - name: xgboostjob
+          - name: xgboost
             image: docker.io/merlintang/xgboost-dist-iris:1.1
             ports:
             - containerPort: 9991
