+++ pkg/controller.v1/mxnet/mxjob_controller.go
@@ -19,6 +19,8 @@ import (
 	"fmt"
 	"time"
 
+	"reflect"
+
 	commonv1 "github.com/kubeflow/common/pkg/apis/common/v1"
 	"github.com/kubeflow/common/pkg/controller.v1/common"
 	"github.com/kubeflow/common/pkg/controller.v1/control"
@@ -30,20 +32,16 @@ import (
 	"github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	kubeclientset "k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/record"
-	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/event"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
-	"sigs.k8s.io/controller-runtime/pkg/source"
 	volcanoclient "volcano.sh/apis/pkg/client/clientset/versioned"
 
 	"github.com/go-logr/logr"
@@ -55,26 +53,25 @@ import (
 const (
 	controllerName = "mxnet-operator"
 
-	// labels for pods and servers.
-	mxReplicaTypeLabel  = "mxnet-replica-type"
-	mxReplicaIndexLabel = "mxnet-replica-index"
-	labelGroupName      = "group-name"
-	labelMXJobName      = "mxnet-job-name"
-	labelMXJobRole      = "mxnet-job-role"
+	// mxJobCreatedReason is added in a mxjob when it is created.
+	mxJobCreatedReason = "MXJobCreated"
+	// mxJobSucceededReason is added in a mxjob when it is succeeded.
+	mxJobSucceededReason = "MXJobSucceeded"
+	// mxJobRunningReason is added in a mxjob when it is running.
+	mxJobRunningReason = "MXJobRunning"
+	// mxJobFailedReason is added in a mxjob when it is failed.
+	mxJobFailedReason = "MXJobFailed"
+	// mxJobRestarting is added in a mxjob when it is restarting.
+	mxJobRestartingReason = "MXJobRestarting"
 )
 
 var (
-	// KeyFunc is the short name to DeletionHandlingMetaNamespaceKeyFunc.
-	// IndexerInformer uses a delta queue, therefore for deletes we have to use this
-	// key function but it should be just fine for non delete events.
-	KeyFunc = cache.DeletionHandlingMetaNamespaceKeyFunc
-
+	jobOwnerKey = ".metadata.controller"
 	// DefaultMXControllerConfiguration is the suggested mxnet-operator configuration for production.
 	DefaultMXControllerConfiguration = common.JobControllerConfiguration{
 		ReconcilerSyncLoopPeriod: metav1.Duration{Duration: 15 * time.Second},
 		EnableGangScheduling:     false,
 	}
-
 	// DefaultCleanPodPolicy is the default clean pod policy controller assign the new Job if not exist
 	DefaultCleanPodPolicy = commonv1.CleanPodPolicyNone
 )
@@ -88,10 +85,12 @@ func NewReconciler(mgr manager.Manager) *MXJobReconciler {
 		Log:      log.Log,
 	}
 
+	// Create clients.
 	cfg := mgr.GetConfig()
 	kubeClientSet := kubeclientset.NewForConfigOrDie(cfg)
 	volcanoClientSet := volcanoclient.NewForConfigOrDie(cfg)
 
+	// Initialize common job controller
 	r.JobController = common.JobController{
 		Controller:       r,
 		Expectations:     expectation.NewControllerExpectations(),
@@ -121,7 +120,6 @@ type MXJobReconciler struct {
 //+kubebuilder:rbac:groups=kubeflow.org,resources=mxjobs/finalizers,verbs=update
 //+kubebuilder:rbac:groups="",resources=pods,verbs=get;list;watch;create;update;patch;delete
 //+kubebuilder:rbac:groups="",resources=services,verbs=get;list;watch;create;delete
-
 func (r *MXJobReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	_ = log.FromContext(ctx)
 	logger := r.Log.WithValues(mxjobv1.Singular, req.NamespacedName)
@@ -133,7 +131,7 @@ func (r *MXJobReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl
 		return ctrl.Result{}, client.IgnoreNotFound(err)
 	}
 
-	// mxjobv1.SetDefaults(mxjob)
+	//mxjobv1.SetDefaults_MXJob(mxjob)
 
 	// Check if reconciliation is needed
 	jobKey, err := common.KeyFunc(mxjob)
@@ -150,7 +148,7 @@ func (r *MXJobReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl
 		return ctrl.Result{}, nil
 	}
 
-	// Set default priorities to pytorch job
+	// Set default priorities to mxnet job
 	scheme.Scheme.Default(mxjob)
 
 	// Convert PyTorch.Spec.PyTorchReplicasSpecs to  map[commonv1.ReplicaType]*commonv1.ReplicaSpec
@@ -180,43 +178,63 @@ func (r *MXJobReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl
 
 // SetupWithManager sets up the controller with the Manager.
 func (r *MXJobReconciler) SetupWithManager(mgr ctrl.Manager) error {
-	// Create a new Controller
-	c, err := controller.New(controllerName, mgr, controller.Options{Reconciler: r})
-	if err != nil {
+	// setup FieldIndexer to inform the manager that this controller owns pods and services,
+	// so that it will automatically call Reconcile on the underlying MXJob when a Pod or Service changes, is deleted, etc.
+	if err := mgr.GetFieldIndexer().IndexField(context.Background(), &corev1.Pod{}, jobOwnerKey, func(rawObj client.Object) []string {
+		pod := rawObj.(*corev1.Pod)
+		owner := metav1.GetControllerOf(pod)
+		if owner == nil {
+			return nil
+		}
+
+		// Make sure owner is XGBoostJob Controller.
+		if owner.APIVersion != r.GetAPIGroupVersion().Version || owner.Kind != r.GetAPIGroupVersionKind().Kind {
+			return nil
+		}
+
+		return []string{owner.Name}
+	}); err != nil {
 		return err
 	}
 
-	// Watch for changes to PyTorchJob
-	err = c.Watch(&source.Kind{Type: &mxjobv1.MXJob{}}, &handler.EnqueueRequestForObject{},
-		predicate.Funcs{CreateFunc: onOwnerCreateFunc()},
-	)
-	if err != nil {
+	if err := mgr.GetFieldIndexer().IndexField(context.Background(), &corev1.Service{}, jobOwnerKey, func(rawObj client.Object) []string {
+		svc := rawObj.(*corev1.Service)
+		owner := metav1.GetControllerOf(svc)
+		if owner == nil {
+			return nil
+		}
+
+		if owner.APIVersion != r.GetAPIGroupVersion().Version || owner.Kind != r.GetAPIGroupVersionKind().Kind {
+			return nil
+		}
+
+		return []string{owner.Name}
+	}); err != nil {
 		return err
 	}
 
-	//inject watching for pytorchjob related pod
-	err = c.Watch(&source.Kind{Type: &corev1.Pod{}}, &handler.EnqueueRequestForOwner{
-		IsController: true,
-		OwnerType:    &mxjobv1.MXJob{},
-	},
-		predicate.Funcs{CreateFunc: util.OnDependentCreateFunc(r.Expectations), DeleteFunc: util.OnDependentDeleteFunc(r.Expectations)},
-	)
-	if err != nil {
-		return err
-	}
+	return ctrl.NewControllerManagedBy(mgr).
+		For(&mxjobv1.MXJob{}).
+		Owns(&corev1.Pod{}).
+		Owns(&corev1.Service{}).
+		Complete(r)
+}
 
-	//inject watching for xgboostjob related service
-	err = c.Watch(&source.Kind{Type: &corev1.Service{}}, &handler.EnqueueRequestForOwner{
-		IsController: true,
-		OwnerType:    &mxjobv1.MXJob{},
-	},
-		&predicate.Funcs{CreateFunc: util.OnDependentCreateFunc(r.Expectations), DeleteFunc: util.OnDependentDeleteFunc(r.Expectations)},
-	)
-	if err != nil {
-		return err
-	}
+// Below is ControllerInterface's implementation
+func (r *MXJobReconciler) ControllerName() string {
+	return controllerName
+}
 
-	return nil
+func (r *MXJobReconciler) GetAPIGroupVersionKind() schema.GroupVersionKind {
+	return mxjobv1.GroupVersion.WithKind(mxjobv1.Kind)
+}
+
+func (r *MXJobReconciler) GetAPIGroupVersion() schema.GroupVersion {
+	return mxjobv1.GroupVersion
+}
+
+func (r *MXJobReconciler) GetGroupNameLabelValue() string {
+	return mxjobv1.GroupVersion.Group
 }
 
 func (r *MXJobReconciler) GetJobFromInformerCache(namespace, name string) (metav1.Object, error) {
@@ -236,7 +254,7 @@ func (r *MXJobReconciler) GetJobFromInformerCache(namespace, name string) (metav
 func (r *MXJobReconciler) GetJobFromAPIClient(namespace, name string) (metav1.Object, error) {
 	job := &mxjobv1.MXJob{}
 
-	clientReader, err := getDelegatingClientFromClient(r.Client)
+	clientReader, err := util.GetDelegatingClientFromClient(r.Client)
 	if err != nil {
 		return nil, err
 	}
@@ -250,6 +268,39 @@ func (r *MXJobReconciler) GetJobFromAPIClient(namespace, name string) (metav1.Ob
 		return nil, err
 	}
 	return job, nil
+}
+
+func (r *MXJobReconciler) GetPodsForJob(obj interface{}) ([]*corev1.Pod, error) {
+	job, err := meta.Accessor(obj)
+	if err != nil {
+		return nil, fmt.Errorf("%v is not a type of MXJob", job)
+	}
+	// List all pods to include those that don't match the selector anymore
+	// but have a ControllerRef pointing to this controller.
+	podlist := &corev1.PodList{}
+	err = r.List(context.Background(), podlist, client.MatchingLabels(r.GenLabels(job.GetName())))
+	if err != nil {
+		return nil, err
+	}
+	return util.ConvertPodList(podlist.Items), nil
+}
+
+func (r *MXJobReconciler) GetServicesForJob(job interface{}) ([]*corev1.Service, error) {
+	mxJob, err := meta.Accessor(job)
+	if err != nil {
+		return nil, fmt.Errorf("%v is not a type of MXJob", mxJob)
+	}
+
+	// List all services to include those that don't match the selector anymore
+	// but have a ControllerRef pointing to this controller.
+	serviceList := &corev1.ServiceList{}
+	err = r.List(context.Background(), serviceList, client.MatchingLabels(r.GenLabels(mxJob.GetName())))
+	if err != nil {
+		return nil, err
+	}
+
+	ret := util.ConvertServiceList(serviceList.Items)
+	return ret, nil
 }
 
 func (r *MXJobReconciler) DeleteJob(job interface{}) error {
@@ -267,24 +318,112 @@ func (r *MXJobReconciler) DeleteJob(job interface{}) error {
 	return nil
 }
 
-// satisfiedExpectations returns true if the required adds/dels for the given mxjob have been observed.
-// Add/del counts are established by the controller at sync time, and updated as controllees are observed by the controller
-// manager.
+func (r *MXJobReconciler) UpdateJobStatus(job interface{}, replicas map[commonv1.ReplicaType]*commonv1.ReplicaSpec, jobStatus *commonv1.JobStatus) error {
+	mxjob, ok := job.(*mxjobv1.MXJob)
+	if !ok {
+		return fmt.Errorf("%v is not a type of MXJob", mxjob)
+	}
 
-func (r *MXJobReconciler) ControllerName() string {
-	return controllerName
+	mxjobKey, err := common.KeyFunc(mxjob)
+	if err != nil {
+		utilruntime.HandleError(fmt.Errorf("couldn't get key for mxjob object %#v: %v", mxjob, err))
+		return err
+	}
+
+	for rtype, spec := range replicas {
+		status := jobStatus.ReplicaStatuses[rtype]
+
+		// Expect to have `replicas - succeeded` pods alive.
+		succeeded := status.Succeeded
+		expected := *(spec.Replicas) - succeeded
+		running := status.Active
+		failed := status.Failed
+
+		r.Log.Info(fmt.Sprintf("MXJob=%s, ReplicaType=%s expected=%d, running=%d, succeeded=%d , failed=%d",
+			mxjob.Name, rtype, expected, running, succeeded, failed))
+
+		if mxjob.Status.StartTime == nil {
+			now := metav1.Now()
+			mxjob.Status.StartTime = &now
+			// enqueue a sync to check if job past ActiveDeadlineSeconds
+			if mxjob.Spec.RunPolicy.ActiveDeadlineSeconds != nil {
+				logrus.Infof("Job with ActiveDeadlineSeconds will sync after %d seconds", *mxjob.Spec.RunPolicy.ActiveDeadlineSeconds)
+				r.WorkQueue.AddAfter(mxjobKey, time.Duration(*mxjob.Spec.RunPolicy.ActiveDeadlineSeconds)*time.Second)
+			}
+		}
+
+		if running > 0 {
+			msg := fmt.Sprintf("MXJob %s is running.", mxjob.Name)
+			err := commonutil.UpdateJobConditions(jobStatus, commonv1.JobRunning, mxJobRunningReason, msg)
+			if err != nil {
+				logrus.Infof("Append mxjob condition error: %v", err)
+				return err
+			}
+		}
+		if expected == 0 {
+			msg := fmt.Sprintf("MXJob %s is successfully completed.", mxjob.Name)
+			r.Recorder.Event(mxjob, corev1.EventTypeNormal, mxJobSucceededReason, msg)
+			if mxjob.Status.CompletionTime == nil {
+				now := metav1.Now()
+				mxjob.Status.CompletionTime = &now
+			}
+			err := commonutil.UpdateJobConditions(jobStatus, commonv1.JobSucceeded, mxJobSucceededReason, msg)
+			if err != nil {
+				logrus.Infof("Append mxjob condition error: %v", err)
+				return err
+			}
+		}
+
+		if failed > 0 {
+			if spec.RestartPolicy == commonv1.RestartPolicyExitCode {
+				msg := fmt.Sprintf("mxjob %s is restarting because %d %s replica(s) failed.", mxjob.Name, failed, rtype)
+				r.Recorder.Event(mxjob, corev1.EventTypeWarning, mxJobRestartingReason, msg)
+				err := commonutil.UpdateJobConditions(jobStatus, commonv1.JobRestarting, mxJobRestartingReason, msg)
+				if err != nil {
+					logrus.Infof("Append job condition error: %v", err)
+					return err
+				}
+			} else {
+				msg := fmt.Sprintf("mxjob %s is failed because %d %s replica(s) failed.", mxjob.Name, failed, rtype)
+				r.Recorder.Event(mxjob, corev1.EventTypeNormal, mxJobFailedReason, msg)
+				if mxjob.Status.CompletionTime == nil {
+					now := metav1.Now()
+					mxjob.Status.CompletionTime = &now
+				}
+				err := commonutil.UpdateJobConditions(jobStatus, commonv1.JobFailed, mxJobFailedReason, msg)
+				if err != nil {
+					logrus.Infof("Append job condition error: %v", err)
+					return err
+				}
+			}
+		}
+	}
+
+	return nil
 }
 
-func (r *MXJobReconciler) GetAPIGroupVersionKind() schema.GroupVersionKind {
-	return mxjobv1.GroupVersion.WithKind(mxjobv1.Kind)
+// UpdateJobStatusInApiServer updates the status of the given MXJob.
+func (r *MXJobReconciler) UpdateJobStatusInApiServer(job interface{}, jobStatus *commonv1.JobStatus) error {
+	mxJob, ok := job.(*mxjobv1.MXJob)
+	if !ok {
+		return fmt.Errorf("%v is not a type of MXJob", mxJob)
+	}
+
+	if !reflect.DeepEqual(&mxJob.Status, jobStatus) {
+		mxJob = mxJob.DeepCopy()
+		mxJob.Status = *jobStatus.DeepCopy()
+	}
+
+	if err := r.Update(context.Background(), mxJob); err != nil {
+		logrus.Error(err, " failed to update MxJob conditions in the API server")
+		return err
+	}
+
+	return nil
 }
 
-func (r *MXJobReconciler) GetAPIGroupVersion() schema.GroupVersion {
-	return mxjobv1.GroupVersion
-}
-
-func (r *MXJobReconciler) GetGroupNameLabelValue() string {
-	return mxjobv1.GroupVersion.Group
+func (r *MXJobReconciler) SetClusterSpec(job interface{}, podTemplate *corev1.PodTemplateSpec, rtype, index string) error {
+	return SetPodEnv(job, podTemplate, rtype, index)
 }
 
 func (r *MXJobReconciler) GetDefaultContainerName() string {
@@ -293,10 +432,6 @@ func (r *MXJobReconciler) GetDefaultContainerName() string {
 
 func (r *MXJobReconciler) GetDefaultContainerPortName() string {
 	return mxjobv1.DefaultPortName
-}
-
-func (r *MXJobReconciler) GetJobRoleKey() string {
-	return labelMXJobRole
 }
 
 func (r *MXJobReconciler) IsMasterRole(replicas map[commonv1.ReplicaType]*commonv1.ReplicaSpec,
