--- pkg/controller.v1/mxnet/service.go
@@ -1,26 +1,0 @@
-package mxnet
-
-import (
-	"context"
-	"fmt"
-	"github.com/kubeflow/tf-operator/pkg/common/util"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/meta"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-func (r *MXJobReconciler) GetServicesForJob(job interface{}) ([]*corev1.Service, error) {
-	mxJob, err := meta.Accessor(job)
-	if err != nil {
-		return nil, fmt.Errorf("%v is not a type of MXJob", mxJob)
-	}
-
-	// List all services to include those that don't match the selector anymore
-	// but have a ControllerRef pointing to this controller.
-	serviceList := &corev1.ServiceList{}
-	err = r.List(context.Background(), serviceList, client.MatchingLabels(r.GenLabels(mxJob.GetName())))
-	if err != nil {
-		return nil, err
-	}
-	return util.ConvertServiceList(serviceList.Items), nil
-}
