--- pkg/controller.v1/pytorch/service.go
@@ -1,48 +1,0 @@
-package controllers
-
-import (
-	"context"
-
-	pytorchv1 "github.com/kubeflow/tf-operator/pkg/apis/pytorch/v1"
-	"github.com/kubeflow/tf-operator/pkg/common/util"
-
-	apiv1 "github.com/kubeflow/common/pkg/apis/common/v1"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-func (r *PyTorchJobReconciler) GetServicesForJob(obj interface{}) ([]*corev1.Service, error) {
-	job, err := meta.Accessor(obj)
-	if err != nil {
-		return nil, err
-	}
-
-	// List all pods to include those that don't match the selector anymore
-	// but have a ControllerRef pointing to this controller.
-	serviceList := &corev1.ServiceList{}
-	err = r.List(context.Background(), serviceList, client.MatchingLabels(r.GenLabels(job.GetName())))
-	if err != nil {
-		return nil, err
-	}
-
-	ret := util.ConvertServiceList(serviceList.Items)
-
-	return ret, nil
-}
-
-// ReconcileServices checks and updates services for each given ReplicaSpec.
-// It will requeue the job in case of an error while creating/deleting services.
-func (r *PyTorchJobReconciler) ReconcileServices(
-	job metav1.Object,
-	services []*corev1.Service,
-	rtype apiv1.ReplicaType,
-	spec *apiv1.ReplicaSpec) error {
-
-	if string(rtype) == string(pytorchv1.PyTorchReplicaTypeMaster) {
-		return r.JobController.ReconcileServices(job, services, rtype, spec)
-	}
-
-	return nil
-}
