--- pkg/controller.v1/xgboost/service.go
@@ -1,43 +1,0 @@
-/*
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package xgboost
-
-import (
-	"context"
-	"fmt"
-	"github.com/kubeflow/tf-operator/pkg/common/util"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/meta"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-// GetServicesForJob returns the services managed by the job. This can be achieved by selecting services using label key "job-name"
-// i.e. all services created by the job will come with label "job-name" = <this_job_name>
-func (r *XGBoostJobReconciler) GetServicesForJob(obj interface{}) ([]*corev1.Service, error) {
-	job, err := meta.Accessor(obj)
-	if err != nil {
-		return nil, fmt.Errorf("%+v is not a type of XGBoostJob", job)
-	}
-	// List all pods to include those that don't match the selector anymore
-	// but have a ControllerRef pointing to this controller.
-	serviceList := &corev1.ServiceList{}
-	err = r.List(context.Background(), serviceList, client.MatchingLabels(r.GenLabels(job.GetName())))
-	if err != nil {
-		return nil, err
-	}
-
-	//TODO support adopting/orphaning
-	ret := util.ConvertServiceList(serviceList.Items)
-
-	return ret, nil
-}
